+++ gateway/gateway.go
@@ -121,8 +121,10 @@ func (i ImmutablePath) IsValid() error {
 var _ path.Path = (*ImmutablePath)(nil)
 
 type CarParams struct {
-	Range *DagByteRange
-	Scope DagScope
+	Range      *DagByteRange
+	Scope      DagScope
+	Order      DagOrder
+	Duplicates *bool
 }
 
 // DagByteRange describes a range request within a UnixFS file. "From" and
@@ -187,6 +189,13 @@ const (
 	DagScopeAll    DagScope = "all"
 	DagScopeEntity DagScope = "entity"
 	DagScopeBlock  DagScope = "block"
+)
+
+type DagOrder string
+
+const (
+	DagOrderDFS     DagOrder = "dfs"
+	DagOrderUnknown DagOrder = "unk"
 )
 
 type ContentPathMetadata struct {
@@ -278,8 +287,11 @@ type IPFSBackend interface {
 	ResolvePath(context.Context, ImmutablePath) (ContentPathMetadata, error)
 
 	// GetCAR returns a CAR file for the given immutable path. It returns an error
-	// if there was an issue before the CAR streaming begins.
-	GetCAR(context.Context, ImmutablePath, CarParams) (ContentPathMetadata, io.ReadCloser, error)
+	// if there was an issue before the CAR streaming begins. If [CarParams.Duplicates]
+	// is nil, or if [CaraParams.Order] is Unknown, the implementer should change it
+	// such that the caller can form the response "Content-Type" header with the most
+	// amount of information.
+	GetCAR(context.Context, ImmutablePath, *CarParams) (ContentPathMetadata, io.ReadCloser, error)
 
 	// IsCached returns whether or not the path exists locally.
 	IsCached(context.Context, path.Path) bool